Add command callback to emit entire scene payload
This allows the scene Player component to know and utilize the entire data payload from the start of the scene.
Useful because it allows the scene to animate through and evolve the scene over time since it knows all future states.
Commands emitted per tick can control the progress of the scene, rather than necessarily _also_ hold the scenes state.